% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Array-class}
\alias{Array-class}
\title{An S4 class to represent a demographic array}
\description{
Differences from ordinary array:
}
\details{
\itemize{
  \item \code{drop} is \code{FALSE} by default
}
}
\section{Slots}{

\describe{
\item{\code{dimtypes}}{Character vector}
}}

