% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGeneric.R,
%   R/Array-generator.R
\docType{class}
\name{Counts-class}
\alias{Counts-class}
\alias{Values-class}
\alias{Array}
\alias{Array,array-method}
\alias{Array,Array-method}
\alias{Counts}
\alias{Values}
\title{Create a demographic array}
\usage{
Array(x, dimtypes = NULL, classif = NULL)

\S4method{Array}{array}(x, dimtypes = NULL, classif = NULL)

\S4method{Array}{Array}(x, dimtypes = NULL, classif = NULL)

Counts(x, dimtypes = NULL, classif = NULL)

Values(x, dimtypes = NULL, classif = NULL)
}
\arguments{
\item{x}{An array, with a complete set of dimnames.}

\item{dimtypes}{A named character specifying dimtypes
to be used by each dimension.}

\item{classif}{A named character vector specifying
classifications to be used by each dimension.}
}
\value{
A \code{\linkS4class{Array}}.
}
\description{
Create an object of class \code{\linkS4class{Array}}.
}
\examples{
x <- array(1:6,
           dim = c(2, 3),
           dimnames = list(sex = c("Female", "Male"),
                           age = c("0-14", "15-64", "65+")))
Array(x)

# specify dimtype
x <- array(1:6,
           dim = c(2, 3),
           dimnames = list(sex = c("Female", "Male"),
                           year = c("1945-1965", "1965-1985", "1985-2000")))
Array(x, dimtypes = c(year = "cohort"))

# specify dimtypes and classification
x <- array(1:6,
           dim = c(2, 3),
           dimnames = list(sex = c("Female", "Male"),
                           year = c("1945-1965", "1965-1985", "1985-2000")))
Array(x,
      dimtypes = c(year = "cohort"),
      classif = c(sex = "Sex", year = "Generation"))
}
