% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Array-methods.R
\name{classif}
\alias{classif}
\alias{classif,Array-method}
\title{Get or set the classifications used by a demographic array}
\usage{
classif(x)

\S4method{classif}{Array}(x)
}
\arguments{
\item{x}{An \code{\linkS4class{Array}}.}
}
\value{
A named vector
}
\description{
Get or set the classifications used by a demographic array
}
\examples{
x <- Array(array(1:6,
                 dim = c(2, 3),
                 dimnames = list(sex = c("Female", "Male"),
                                 year = c("1945-1965", "1965-1985", "1985-2000"))),
           classif = c(sex = "Sex", year = "Generation"))
classif(x)
}
