% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demarray.R
\docType{package}
\name{demarray}
\alias{demarray}
\title{demarray: Demographic arrays.}
\description{
The demarray (pronounced "dem-array") package contains
functions for creating demographic arrays, plus methods
for some basic functions such as print and `[`.
}
