% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGeneric.R,
%   R/DemographicArray-generators.R
\docType{class}
\name{Counts-class}
\alias{Counts-class}
\alias{Values-class}
\alias{DemographicArray}
\alias{DemographicArray,array-method}
\alias{DemographicArray,DemographicArray-method}
\alias{Counts}
\alias{Values}
\title{Create a demographic array}
\usage{
DemographicArray(x, dimtypes = NULL)

\S4method{DemographicArray}{array}(x, dimtypes = NULL)

\S4method{DemographicArray}{DemographicArray}(x, dimtypes = NULL)

Counts(x, dimtypes = NULL)

Values(x, dimtypes = NULL)
}
\arguments{
\item{x}{An array, with a complete set of dimnames.}

\item{dimtypes}{A named character vector specifying dimtypes
to be used by each dimension.}
}
\value{
A \code{\linkS4class{DemographicArray}}.
}
\description{
Create an object of class \code{\linkS4class{DemographicArray}}.
}
\examples{
x <- array(1:6,
           dim = c(2, 3),
           dimnames = list(sex = c("Female", "Male"),
                           age = c("0-14", "15-64", "65+")))
DemographicArray(x)

# specify dimtype
x <- array(1:6,
           dim = c(2, 3),
           dimnames = list(sex = c("Female", "Male"),
                           year = c("1945-1965", "1965-1985", "1985-2000")))
DemographicArray(x, dimtypes = c(year = "cohort"))

# specify dimtypes
x <- array(1:6,
           dim = c(2, 3),
           dimnames = list(sex = c("Female", "Male"),
                           year = c("1945-1965", "1965-1985", "1985-2000")))
DemographicArray(x,
      dimtypes = c(year = "cohort"))
}
