% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Array-methods.R
\name{dimtypes}
\alias{dimtypes}
\alias{dimtype}
\alias{dimtypes,Array-method}
\title{Get or set dimtypes}
\usage{
dimtypes(x)

\S4method{dimtypes}{Array}(x)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{Array}}.}
}
\value{
A named vector
}
\description{
Query or modify the dimtypes of an object.
}
\details{
Every dimension of a  demographic array has a
'dimtype' describing the type of information held by that dimension.
Examples include \code{"age"}, \code{"time"}, and \code{"state"}. See below
for details.

The range of dimtypes are as follows:

\tabular{rll}{ dimtype \tab Description \tab Labels \cr
\code{"age"} \tab Age, as points or intervals \tab <any> \cr
\code{"time"} \tab Time, as points or intervals \cr
\code{"cohort"} \tab Cohort, measured as points or intervals \tab <any> \cr
\code{"attribute"} \tab Any qualitative attribute \tab  <any> \cr
\code{"origin"}, \code{"destination"} \tab Starting and finishing states
\tab <any> \cr
\code{"parent"}, \code{"child"} \tab Parent's state versus child's state
\tab <any> \cr
\code{"iteration"} \tab Simulation number \tab Positive integers \cr
\code{"quantile"} \tab Position in distribution \tab Percentage \cr
\code{"triangle"} \tab Lexis triangle \tab \code{"Lower", "Upper"} \cr
\code{"pool"} \tab Type of migration flow \tab \code{"Ins", "Outs"} \cr
}

\code{"origin"}, \code{"destination"}.  Starting and
finishing values for an attribute that is subject to change, such as country
of residence.  \code{"origin"} and \code{"destination"} dimensions must come
in pairs with a common base name and suffixes \code{"_orig"} and
\code{"_dest"}, each \code{"country_orig"} and \code{"country_dest"}.  There
is no limit to the number of origin-destination pairs that an object may
have.

\code{"parent"}, \code{"child"}.  Parents' attributes versus
those of their children.  Used to model transmission of attributes such as
ethnicity from parents to children. \code{"parent"} and \code{"child"}
dimensions must come in pairs with a common base name and suffixes
\code{"_parent"} and \code{"_child"}, e.g. \code{"ethnicity_parent"} and
\code{"ethnicity_child"}.  There is no limit to the number of parent-child
pairs that an object may have.

\code{"iteration"}.  Iteration number
from a simulation.  An object can have at most one dimension with dimtype
\code{"iteration"}.  If an object has a dimension with dimtype
\code{"iteration"} it may not have a dimension with dimtype
\code{"quantile"}.

\code{"quantile"}.  Sample quantiles, typically
summarizing simulation results.  An object can have at most one dimension
with dimtype \code{"quantile"}.  If an object has a dimension with dimtype
\code{"quantile"} it may not have a dimension with dimtype
\code{"iteration"}.

\code{"triangle"}.  Lexis triangle.

\code{"pool"}. Direction of flow in "pool" model of migration.
}
\examples{
x <- Array(array(1:6,
                 dim = c(2, 3),
                 dimnames = list(sex = c("Female", "Male"),
                                 time = c(2000, 2010, 2020))))
dimtypes(x)
}
