% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_lab.R
\name{infer_lab}
\alias{infer_lab}
\title{Derive a "Labels" object from a set of
labels and a dimtype}
\usage{
infer_lab(labels, dimtype)
}
\arguments{
\item{labels}{A character vector.}

\item{dimtype}{A character vector of length 1.}
}
\value{
An object of class \code{\linkS4class{Labels}}.
}
\description{
Derive an object of class \code{\linkS4class{Labels}}
from a vector of labels and the name of a dimtype.
\code{infer_lab} is not normally called directly
by end-users.
}
\keyword{internal}
