% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\name{is_quant_scale}
\alias{is_quant_scale}
\alias{is_quant_scale,DemographicArray-method}
\title{Whether cross-classifying dimensions have quantitative scales}
\usage{
is_quant_scale(x)

\S4method{is_quant_scale}{DemographicArray}(x)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{DemographicArray}}.
\emph{Add reference to demographic account, once we have implemented
them.}}
}
\value{
A named logical vector.
}
\description{
Check how functions in \code{demarray} and associated packages
will interpret the labels along each dimension. In particular,
check whether the functions will interpret the labels as representing
a quantitative scale, such as age or calendar time.
}
\details{
To change the labels (and hence, potentially, to switch
between quantitative and qualitative scales), use
function \code{\link[base]{dimnames}}.
}
\examples{
x <- Counts(array(1:8,
                  dim = c(2, 2, 2),
                  dimnames = list(sex = c("Female", "Male"),
                                  age = c("0-39", "40+"),
                                  quarter = c("2015 Q1", "2015 Q3"))))
x
is_quant_scale(x)
dimnames(x)$age <- c("Young", "Old")
x
is_quant_scale(x)
}
\seealso{
dimtypes
}
