% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Labels-class}
\alias{Labels-class}
\alias{LabCategories-class}
\alias{LabTriangles-class}
\alias{LabPool-class}
\alias{LabQuantiles-class}
\alias{LabIntegers-class}
\alias{LabGroupedInt-class}
\alias{LabGroupedIntEnumerations-class}
\alias{LabGroupedIntEndpoints-class}
\alias{LabCalendar-class}
\alias{LabCalendarQuarters-class}
\alias{LabCalendarMonths-class}
\alias{LabDurations-class}
\alias{LabDurationsQuarters-class}
\alias{LabDurationsMonths-class}
\title{S4 classes to represent information in labels}
\description{
S4 classes to hold information extracted from labels.
End users would not normally interact directly with
these classes.
}
\section{Slots}{

\describe{
\item{\code{labels}}{Character vector holding labels.}

\item{\code{int_min}}{Integer scalar. Lowest value for integer labels.}

\item{\code{int_max}}{Integer scalar. Highest value for integer labels.}

\item{\code{breaks}}{Integers or dates. Boundaries between intervals.}

\item{\code{break_min}}{Integer scalar or date. Lower boundary for intervals.}

\item{\code{break_max}}{Integer scalar or date. Upper boundary for intervals.}

\item{\code{open_first}}{Logical flag. Whether there is an interval
open on the left.}

\item{\code{open_last}}{Logical flag. Whether there is an interval
open on the right.}

\item{\code{include_na}}{Logical. Whether labels include an
\code{NA}.}
}}

\keyword{internal}
