% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/array-methods.R
\name{limits}
\alias{limits}
\alias{limits,array-method}
\title{Get first and last labelsl used by each dimension.}
\usage{
limits(x)

\S4method{limits}{array}(x)
}
\arguments{
\item{x}{An object of class \code{\link[base]{array}},
\code{\linkS4class{DemographicArray}}, or a demographic account.}
}
\value{
A data frame.
}
\description{
Get the first and last categories used by each dimension
of an array or demographic account. These categories can
be useful in understanding the structure of the array
or account. When used with an account, \code{limits}
returns the values for the population series.
}
\examples{
x <- Counts(array(1:8,
                  dim = c(2, 4),
                  dimnames = list(sex = c("Female", "Male"),
                                  age = c("0-14", "15-39", "40-64", "65+"))))
x
limits(x)
}
